[chore] change default LOCAL parameter to false in run-app.sh
In run-app.sh, the LOCAL flag was initialized to true by default which causes the script to always short-circuit in local mode; update the default to false so cluster runs are executed as intended.
Modified the default value of LOCAL from true to false in the run-app.sh script configuration  changes the default behavior to run in non-local mode.